% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/basic.R
\name{HSDS_dataset}
\alias{HSDS_dataset}
\title{create HSDS dataset reference amenable to bracket subsetting}
\usage{
HSDS_dataset(txsrc)
}
\arguments{
\item{instance}{of \code{\link{HSDS_source-class}}}
}
\description{
create HSDS dataset reference amenable to bracket subsetting
}
