% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/basic.R
\docType{class}
\name{HSDS_source-class}
\alias{HSDS_source-class}
\alias{show,HSDS_source-method}
\alias{HSDS_source}
\title{class for managing information about an HSDS 'host'}
\usage{
\S4method{show}{HSDS_source}(object)

HSDS_source(serverPort, hostPath)
}
\arguments{
\item{serverPort}{character(1) fully qualified URL and port number for HSDS instance}

\item{hostPath}{character(1) string with resource name}
}
\description{
class for managing information about an HSDS 'host'

display HSDS_source instance

constructor for HSDS_source
}
\examples{
hostPath = "/home/john/tenx_full.h5"
serverPort = "http://149.165.156.174:5101/"
HSDS_source(serverPort=serverPort, hostPath=hostPath)
}
